client can make a reservation if available, company update status function added, SQL constraints added, Names changed according request

+++ src/main/java/yt/company/carwash/models/Company.java
@@ -16,8 +16,7 @@ public class Company extends BaseModel {
     private String name;
     private String address;
     private String phone;
-    private short capacity;
-
+    private int capacity;
     @ManyToMany
     private List<VehicleType> vehicleTypes;
     @OneToMany
